requirements.txt
streamlit
streamlit-option-menu
pandas
numpy
matplotlib
seaborn
scikit-learn
lasio
striplog
openpyxl
Pillow
